team-service/requirements.txt
asyncpg==0.23.0
fastapi==0.65.3
requests==2.25.1
uvicorn==0.14.0
tortoise-orm==0.17.4
databases[postgresql]==0.2.6
black==21.6b0
flake8==3.9.2
isort==5.9.1
pytest==6.2.4
pytest-cov==2.12.1
pytest-xdist==2.3.0